Moving targets to their own package
