+++ libs/SalesforceHybrid/src/com/salesforce/androidsdk/phonegap/plugin/SmartSyncPlugin.java
@@ -31,10 +31,10 @@ import android.util.Log;
 import com.salesforce.androidsdk.smartstore.store.SmartStore;
 import com.salesforce.androidsdk.smartsync.manager.SyncManager;
 import com.salesforce.androidsdk.smartsync.manager.SyncManager.SyncUpdateCallback;
-import com.salesforce.androidsdk.smartsync.util.SyncDownTarget;
+import com.salesforce.androidsdk.smartsync.target.SyncDownTarget;
 import com.salesforce.androidsdk.smartsync.util.SyncOptions;
 import com.salesforce.androidsdk.smartsync.util.SyncState;
-import com.salesforce.androidsdk.smartsync.util.SyncUpTarget;
+import com.salesforce.androidsdk.smartsync.target.SyncUpTarget;
 
 import org.apache.cordova.CallbackContext;
 import org.json.JSONArray;
