+++ libs/SmartSync/src/com/salesforce/androidsdk/smartsync/util/SyncState.java
@@ -28,6 +28,9 @@ package com.salesforce.androidsdk.smartsync.util;
 
 import com.salesforce.androidsdk.smartstore.store.IndexSpec;
 import com.salesforce.androidsdk.smartstore.store.SmartStore;
+import com.salesforce.androidsdk.smartsync.target.SyncDownTarget;
+import com.salesforce.androidsdk.smartsync.target.SyncTarget;
+import com.salesforce.androidsdk.smartsync.target.SyncUpTarget;
 
 import org.json.JSONArray;
 import org.json.JSONException;
